to save

+++ src/cgi_Response.cpp
@@ -6,7 +6,7 @@
 /*   By: iantar <iantar@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 23:03:14 by iantar            #+#    #+#             */
-/*   Updated: 2024/03/24 02:04:37 by iantar           ###   ########.fr       */
+/*   Updated: 2024/03/24 02:05:26 by iantar           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ std::string Response::getScriptName()
 
 void Response::print_CGI_env(char **env)
 {
-	for (size_t i = 0; i < 8; i++)
+	for (int i = 0; i < 8; i++)
 	{
 		std::cout << env[i] << std::endl;
 	}
